Move AdminChronografPage to /admin/chronograf
