+++ ui/src/admin/containers/chronograf/AdminChronografPage.js
