+++ ui/src/admin/index.js
@@ -1,4 +1,4 @@
 import AdminInfluxDBPage from './containers/AdminInfluxDBPage'
-import AdminChronografPage from './containers/AdminChronografPage'
+import AdminChronografPage from './containers/chronograf/AdminChronografPage'
 
 export {AdminChronografPage, AdminInfluxDBPage}
